Move gui things into a gui package
